DatabaseCreator/requirements.txt
elasticsearch==8.12.2
more_itertools==10.7.0
mysql_connector_repackaged==0.3.1
python-dotenv==1.1.0
python_dateutil==2.9.0.post0
Requests==2.32.3
